requirements.txt
aiogram==3.24.0
supabase==2.13.0
thefuzz==0.22.1
python-levenshtein==0.26.1
python-dotenv==1.0.1
openai>=1.0.0
websockets>=13.0
